webapp/backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv 